[AiBundle][McpBundle] Upgrade bundle structure
