+++ src/ai-bundle/templates/data_collector.html.twig
