+++ src/mcp-bundle/config/routes.php
