+++ src/mcp-bundle/config/services.php
